fix(execution): handled all the exit values when execve fails
